refactor: move SYSTEM_HANDLE_INFORMATION_EX to Windows.Win32
